pyproject.toml
[dependency-groups]
dev = [
  "pre-commit>=4.5.1",
  "pytest-cov>=7.0.0",
  "pytest-env>=1.2.0",
  "pytest>=9.0.2",
  "ruff>=0.14.10",
]

[project]
authors = [{email = "marcelpaluch@proton.me", name = "marcelpaluch"}]
dependencies = [
  "alembic>=1.17.2",
  "basedpyright>=1.36.2",
  "fastapi[standard]>=0.127.1",
  "loguru>=0.7.3",
  "psycopg2-binary>=2.9.11",
  "pydantic-settings>=2.12.0",
  "pydantic>=2.12.5",
  "python-slugify>=8.0.4",
  "rich>=14.2.0",
  "sqlmodel>=0.0.30",
]
description = "Add your description here"
name = "repository-infrastructure-example"
readme = "README.md"
requires-python = ">=3.14"
version = "0.1.0"

[tool.basedpyright]
exclude = ["./.venv/*", "./alembic/versions/*"]
executionEnvironments = [{root = "src"}]
reportMissingTypeStubs = false
reportUnnecessaryTypeIgnoreComment = true
typeCheckingMode = "strict"
venv = ".venv"

[tool.pytest.ini_options]
addopts = "-ra --durations=10 --cov=src --cov-report=term-missing"

[tool.uv]
package = true
